requirements.txt
asgiref==3.7.2
boto3==1.28.59
botocore==1.31.59
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
Django==4.2.5
django-filter==23.3
django-storages==1.14.1
fonttools==4.43.0
gunicorn==21.2.0
html5lib==1.1
jmespath==1.0.1
kaleido==0.2.1
mysqlclient==2.2.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
tinycss2==1.2.1
tzdata==2023.3
urllib3==1.26.17
weasyprint==60.1
webencodings==0.5.1
whitenoise==6.5.0
zopfli==0.2.3
